refactor: move state types to types.go
